split up Lightning modules to separate files
